% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.nl.r
\name{invlogit}
\alias{invlogit}
\title{Inverse logit transformation}
\usage{
invlogit(x)
}
\arguments{
\item{x}{A real number}
}
\value{
A number between \eqn{0} and \eqn{1}.
}
\description{
Inverse transformation of the \code{logit} function.
}
