% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAmevt-package.R
\docType{data}
\name{Leeds}
\alias{Leeds}
\title{Tri-variate \sQuote{angular} data set approximately distributed according to a multivariate extremes angular distribution}
\format{
A \eqn{100*3} - matrix.
}
\description{
The data set is constructed from coordinates (columns) \eqn{1,2,3} of \code{\link{frechetdat}}.
It contains 100 angular points corresponding to the tri-variate vectors \eqn{V=(X,Y,Z)} with largest \eqn{L^1}{L1} norm (\eqn{||V||=X+Y+Z}). The angular points are obtained by \sQuote{normalizing}: \emph{e.g.},
\eqn{x=X/||V||}. Thus, 
each row in \code{Leeds} is a point on the two-dimensional simplex : \eqn{x+y+z=1}.
}
\references{
COOLEY, D., DAVIS, R. and NAVEAU, P. (2010). The pairwise beta distribution: A flexible parametric multivariate model for extremes. \emph{Journal of Multivariate Analysis 101, 2103-2117}

RESNICK, S. (1987). Extreme values, regular variation, and point processes, \emph{Applied Probability. A, vol. 4, 
Series of the Applied Probability Trust. Springer-Verlag, New York}.
}
\keyword{datasets}
