% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAmevt-package.R
\docType{data}
\name{nl.Hpar}
\alias{nl.Hpar}
\title{Default hyper-parameters for the NL model.}
\format{
A list of four parameters: \describe{
\item{mean.alpha, sd.alpha}{%
Mean and standard deviation of the normal prior distribution for the logit-transformed global dependence parameter \eqn{alpha} .
Default to \eqn{0, 3}.
}
\item{mean.beta, sd.beta}{Idem for the pairwise dependence parameters.
}
}
}
\description{
The logit-transformed parameters for the NL model are \emph{a priori}
Gaussian. The list has the same format as \code{\link{pb.Hpar}}.
}
\keyword{dataset}
