\docType{data}
\name{nl.MCpar}
\alias{nl.MCpar}
\title{Default MC MC tuning parameter for the Nested Asymmetric logistic model.}
\format{A list made of a  single element: \code{sd}. The standard deviation of the normal proposition kernel centered at the (logit-transformed)
current state. Default to \eqn{0.35}.}
\description{
  The proposals (on the logit-scale) are Gaussian, centered
  aroud the current value.
}
\keyword{dataset}
\keyword{datasets}

