% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAmevt-package.R
\docType{data}
\name{winterdat}
\alias{winterdat}
\title{Five-dimensional air quality dataset recorded in Leeds(U.K.), during five winter seasons.}
\format{A \eqn{590*5} matrix.}
\source{
\url{http://www.airquality.co.uk}
}
\description{
Contains 590 daily maxima of five air pollutants
(respectively PM10, N0, NO2, 03, S02) recorded in Leeds (U.K.)
during five winter seasons (1994-1998, November-February included). Contains NA's.
}
\keyword{datasets}
