% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMAmevt-package.R
\docType{package}
\name{BMAmevt-package}
\alias{BMAmevt-package}
\alias{BMAmevt}
\title{Bayesian Model Averaging for Multivariate Extremes}
\description{
Toolkit for Bayesian estimation of the dependence structure
in Multivariate Extreme Value parametric models, with possible use of Bayesian model Averaging techniques
Includes a Generic MCMC sampler. Estimation of the marginal
distributions is a prerequisite, \emph{e.g.} using one of the
packages
\code{ismev}, \code{evd}, \code{evdbayes} or \code{POT}. This package handles data sets which are assumed
to be marginally unit-Frechet distributed.
}
\seealso{
\code{evdbayes}
}
\author{
Anne Sabourin
}
