\docType{package}
\name{BMAmevt-package}
\alias{BMAmevt}
\alias{BMAmevt-package}
\title{Bayesian Model Averaging for  Multivariate Extremes}
\description{
  Toolkit for Bayesian estimation of the dependence
  structure in Multivariate Extreme Value parametric
  models, with possible use of Bayesian model Averaging
  techniques
}
\details{
  \tabular{ll}{ Package: \tab BMAmevt \cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2012-03-1\cr
  License: \tab GPL-2\cr LazyData: \tab no\cr } Includes a
  Generic MC MC sampler. Estimation of the marginal
  distributions is a prerequisite, \emph{e.g.} using one of
  the packages \code{ismev}, \code{evd}, \code{evdbayes} or
  \code{POT}. This package handles data sets which are
  assumed to be marginally unit-Frechet distributed.
}
\author{
  Anne Sabourin \email{sabourin@math.univ-lyon1.fr}
}
\seealso{
  \code{evdbayes}
}
\keyword{htest}
\keyword{models}
\keyword{multivariate}
\keyword{package}

