\docType{data}
\name{Leeds.frechet}
\alias{Leeds.frechet}
\title{Multivariate data set with margins following unit Frechet
distribution.}
\format{A \eqn{590*5} - matrix:}
\source{
  \url{http://www.airquality.co.uk}
}
\description{
  The data set contains 590 (transformed) daily maxima of
  five air pollutants recorded in Leeds (U.K.)  during five
  winter seasons (1994-1998).  Contains NA's. Marginal
  transformation to unit Frechet was performed by Cooley
  \emph{et.al.} (see reference below).
}
\references{
  COOLEY, D., DAVIS , R. and NAVEAU , P. (2010). The
  pairwise beta distribution: A flexible parametric
  multivariate model for extremes. \emph{Journal of
  Multivariate Analysis 101, 2103-2117}.
}
\keyword{datasets}

